scraper/requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
dnspython==2.5.0
idna==3.6
mysql-connector-python==8.3.0
pymongo==4.6.1
requests==2.31.0
setuptools==69.0.3
soupsieve==2.5
urllib3==2.1.0
wheel==0.42.0
